Project refactoring with jsconfig file
