--- src/services/hnApi.js
@@ -1,15 +1,0 @@
-import axios from 'axios';
-
-import selectFields from '../utils/selectFields';
-
-export const baseUrl = 'https://hacker-news.firebaseio.com/v0/';
-export const jobsUrl = `${baseUrl}jobstories.json`;
-export const storyUrl = `${baseUrl}item/`;
-
-export const getStory = async (storyId) => {
-  const res = await axios
-    .get(`${storyUrl + storyId}.json`)
-    .then(({ data }) => data && selectFields(data));
-
-  return res;
-};
